The Impossible Bootstrap: A Tale of Commits
A Brief Introduction
--------------------

This is the first story-style commit for my pet little language, Alice.
Alice is a stack-based language heavily inspired by Forth. The main goal
of this project is to give me a chance to explore the process of writing
languages. My end goal is to fully bootstrap Alice on raw hardware like
the Raspberry Pi or Arduino.

I'm going to be documenting the process of Alice's creation through long
blog-style commit messages like this one. Hopefully this process will
prove to be interesting enough to attract some readers!

NOTE: This language started as a quickly hacked-together
proof-of-concept, and the code is a bit messy as a result. I've
documented it as best as I can, but it is in dire need of a refactor.
Also, the terminology and execution model is in a state of flux.
This blog post is part of the brainstorming process to evolve the
execution model and terminology, so it may differ slightly from the
code. This is probably frustrating as a reader, but don't worry -- it'll
all settle down after a few commits (I promise!).

Alice
-----

As mentioned above, Alice is a stack-based language inspired by Forth.
It strived to be as simple as possible, but to allow for an expressive
and flexible language. In this vein, the execution model is a bit more
complicated to allow for some more powerful metaprogramming features
like infix and postfix operators.

Words and Blocks

The execution model of Alice is based around lines and words. Words are
separated by whitespace and are used as keys to look up code blocks in
the available dictionaries. Each word can have both a "parse" block and
a "execute" block, which are executed during the parse and execution
stages, respectively.

Execution Model

The execution model operates on each line in the program sequentially,
and is broken up into a Parsing and an Executing stage.

In the parsing stage, words are read from left to right, recursively.
Each word is used to look up a parse block in the available dictionaries.
If one is found, a reference to the parser and the remaining line are
pushed onto the _values stack, and the parse block is called. Parse
blocks can reorder the _words stack, modify the current line, or supply
an alternate parser. This allows parse blocks to perform extensive
metaprogramming.

Once the end of the line is released, the Executing phase
begins. During this phase, words are popped off of the _words stack
one-by-one. The word is used to look up a code block in the available
dictionaries. If a block is found, it is either placed onto the _values
stack (if its "defer" property is true), or executed immediately. If no
block is found, the word is pushed onto the _values stack as a string.

Ending Notes
------------

Hopefully this commit does a good job of documenting the execution model
of Alice. As the language starts growing proper documentation, I'll use
these commit messages to explore the design and future roadmap of Alice.

That's it for now, everyone! Hopefully you enjoyed reading about my
language musings. Feel free to send me an email at sean@v13inc.com if
you have any questions or feedback! If you have somehow travelled to a
future world where there's a comment box below, don't hesitate to leave
your thoughts!

- Sean