Refactor FTW
I'll expand on most of the changes in the next commit message, but this
turned out to be a very successful refactor. I've managed to clean up a
LOT of the code, and got everything down to 400 lines! The parser is
still a bit buggy, and everything needs to be properly namespaces, but
its close to perfect (heh)!